updates to alloc isAccessible func.
