+++ src/umpire/alloc/CudaMallocAllocator.hpp
@@ -10,6 +10,7 @@
 #include <cuda_runtime_api.h>
 
 #include "umpire/util/Macros.hpp"
+#include "umpire/util/Platform.hpp"
 
 namespace umpire {
 namespace alloc {
@@ -56,6 +57,33 @@ struct CudaMallocAllocator {
                                       << cudaGetErrorString(error));
     }
   }
+
+  bool isHostPageable()
+  {
+    int pageableMem = 0;
+    int cdev = 0;
+    cudaGetDevice(&cdev);
+
+    //Device supports coherently accessing pageable memory
+    //without calling cudaHostRegister on it
+    cudaDeviceGetAttribute(&pageableMem,
+              cudaDevAttrPageableMemoryAccess, cdev);
+    if(pageableMem)
+      return true;
+    else
+      return false;
+  }
+
+  bool isAccessible(umpire::Platform p)
+  {
+    if(p == umpire::Platform::cuda)
+      return isHostPageable();
+    else if(p == umpire::Platform::host)
+      return true;
+    else
+      return false;
+  }
+
 };
 
 } // end of namespace alloc
